fix: break up programtable and fix scrolling
