+++ client/src/components/dashboard/ProgramTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProgramTable';
